Move Puppet module specific code into dedicated class
